Make batching ec fiber-aware
